refactor: outsource getDriverOptionsFromCall to dedicated shared function
